# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Pokedex api with golang.",
        "title": "Swagger Pokedex API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/v1/pokemon/{name}": {
            "get": {
                "description": "Get pokemon's information",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pokemon's Info"
                ],
                "summary": "Get pokemon's information",
                "parameters": [
                    {
                        "type": "string",
                        "description": "pokemon name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.PokemonInfo"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/pokemons": {
            "get": {
                "description": "Get a list of pokemons",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "List of Pokemons"
                ],
                "summary": "Get a list of pokemons",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Number of pokemons",
                        "name": "limit",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Number of pokemons to exclude",
                        "name": "offset",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Pokemons"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.Abilities": {
            "type": "object",
            "properties": {
                "ability": {
                    "$ref": "#/definitions/models.Ability"
                }
            }
        },
        "models.Ability": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
        "models.Forms": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
        "models.GameIndices": {
            "type": "object",
            "properties": {
                "game_index": {
                    "type": "integer"
                },
                "version": {
                    "$ref": "#/definitions/models.Version"
                }
            }
        },
        "models.Move": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
        "models.MoveLearnMethod": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
        "models.Moves": {
            "type": "object",
            "properties": {
                "move": {
                    "$ref": "#/definitions/models.Move"
                },
                "version_group_details": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.VersionGroupDetails"
                    }
                }
            }
        },
        "models.PokemonInfo": {
            "type": "object",
            "properties": {
                "abilities": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Abilities"
                    }
                },
                "base_experience": {
                    "type": "integer"
                },
                "forms": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Forms"
                    }
                },
                "game_indices": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.GameIndices"
                    }
                },
                "height": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "moves": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Moves"
                    }
                },
                "name": {
                    "type": "string"
                },
                "order": {
                    "type": "integer"
                },
                "species": {
                    "$ref": "#/definitions/models.Species"
                },
                "sprites": {
                    "$ref": "#/definitions/models.Sprites"
                },
                "stats": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Stats"
                    }
                },
                "types": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Types"
                    }
                },
                "weight": {
                    "type": "integer"
                }
            }
        },
        "models.Pokemons": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "sprites": {
                    "$ref": "#/definitions/models.Sprites"
                }
            }
        },
        "models.Species": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
        "models.Sprites": {
            "type": "object",
            "properties": {
                "back_default": {
                    "type": "string"
                },
                "front_default": {
                    "type": "string"
                }
            }
        },
        "models.Stat": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
        "models.Stats": {
            "type": "object",
            "properties": {
                "base_stat": {
                    "type": "integer"
                },
                "effort": {
                    "type": "integer"
                },
                "stat": {
                    "$ref": "#/definitions/models.Stat"
                }
            }
        },
        "models.Type": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
        "models.Types": {
            "type": "object",
            "properties": {
                "type": {
                    "$ref": "#/definitions/models.Type"
                }
            }
        },
        "models.Version": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
        "models.VersionGroup": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
        "models.VersionGroupDetails": {
            "type": "object",
            "properties": {
                "level_learned_at": {
                    "type": "integer"
                },
                "move_learn_method": {
                    "$ref": "#/definitions/models.MoveLearnMethod"
                },
                "version_group": {
                    "$ref": "#/definitions/models.VersionGroup"
                }
            }
        }
    }
}